Update 1.0.6 AddedShooting

+++ src/model/LaserBullet.java
@@ -2,7 +2,7 @@ package model;
 
 public class LaserBullet extends OnScreenItems implements IBullet {
 
-    public LaserBullet(int itemCoordX, int itemCoordY, boolean isFacingPlayer) {
+    public LaserBullet(double itemCoordX, double itemCoordY, boolean isFacingPlayer) {
         super(Constants.laserBulletURL, itemCoordX, itemCoordY, Constants.laserBulletHeight, Constants.laserBulletWidth, Constants.laserBulletMovementSpeed, isFacingPlayer);
     }
