+++ src/model/PlayerShip.java
@@ -3,7 +3,7 @@ package model;
 public class PlayerShip extends Ship {
 
     public PlayerShip() {
-        super(Constants.playerShipURL, (int)Constants.playerShipStartPosX, (int)Constants.playerShipStartPosY, Constants.playerShipHeight, Constants.playerShipWidth, Constants.playerShipMovementSpeed, false);
+        super(Constants.playerShipURL, Constants.playerShipStartPosX, Constants.playerShipStartPosY, Constants.playerShipHeight, Constants.playerShipWidth, Constants.playerShipMovementSpeed, false);
         setWeapons(new LaserGun(false, (int)Constants.laserBulletShootInterval, this));
     }
